+++ include/OpenCore/Animation/MoveAnimation.hpp
@@ -5,7 +5,7 @@
 class MoveAnimation : public IAnimation {
 public:
     // 构造函数，传入起始和结束坐标、持续时间、是否循环
-    MoveAnimation(int startX, int startY, int endX, int endY, float duration, bool isLooping);
+    MoveAnimation(int16_t startX, int16_t startY, int16_t endX, int16_t endY, float duration, bool isLooping);
 
     // 更新动画状态
     void onUpdate(float totalTime, AnimationState& state) override;
@@ -18,7 +18,7 @@ public:
 
 private:
     float startTime = 0.0f; // 起始时间
-    int startX, startY, endX, endY; // 起始和结束坐标
+    int16_t startX, startY, endX, endY; // 起始和结束坐标
     float duration;                 // 持续时间
     bool isLooping;                 // 是否循环
     bool finished;                  // 是否结束
